pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "co-occurrence-library"
version = "0.1.0"
description = "Análisis bibliométrico multivariado en turismo — co-ocurrencia, comunidades y grafos"
authors = [
  { name = "Eduardo Cepeda, Ph.D." }
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.13"
dependencies = [
    "loguru",
    "networkx>=3.4",
    "openpyxl",
    "pandas>=2.2",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "typer",
]

[project.scripts]
co-occurrence = "co_occurrence.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.9",
    "plotly>=5.24",
    "pyvis>=0.3.2",
]
dimred = [
    "prince>=0.13",
    "scikit-learn>=1.5",
    "umap-learn>=0.5",
]
communities = [
    "igraph>=0.11",
    "leidenalg>=0.10",
    "python-louvain>=0.16",
]
topics = [
    "bertopic>=0.16",
]
publish = [
    "itables>=2.2",
    "jupyter",
]
all = [
    "co-occurrence-library[viz,dimred,communities,topics,publish]",
]
dev = [
    "co-occurrence-library[all]",
    "bandit>=1.8.3",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "flake8-pyproject>=1.2.4",
    "isort>=6.0.1",
    "mypy>=1.5.1",
    "pre-commit>=3.7.0",
    "pytest>=8.4.1",
]

# --- Linters / formatters / mypy: single source of truth (pre-commit and direct runs use this) ---

[tool.black]
line-length = 99
target-version = ["py313"]
include = '\.pyi?$'
extend-exclude = '''
^/(
  (\.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|venv|\.env|env|\.vscode|
   _build|buck-out|build|dist|notebooks|references|data|saagie|cicd_saagie_tool|code)/
)
'''

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".venv", "venv", "env", "notebooks", "references", "data"]

[tool.flake8]
max-line-length = 99
ignore = ["E501", "W503", "E203", "DJ01"]
per-file-ignores = []
exclude = [".venv", "venv", ".env", "env", "build", "dist", "notebooks", "references", "data"]

[tool.bandit]
skip = ["B101"]
exclude_dirs = ["tests", ".venv", "venv", ".env", "env", "build", "dist", "notebooks", "references", "data"]

[tool.vulture]
min_confidence = 80
exclude = [".venv", "venv", ".env", "env", "build", "dist", "notebooks", "references", "data"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
ignore_missing_imports = true
implicit_reexport = true
follow_imports = "skip"
exclude = [
    ".venv/Lib/site-packages/*",
    "venv/Lib/site-packages/*",
    ".venv/*",
    "venv/*",
    ".env/*",
    "env/*",
]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_decorators = false
disallow_untyped_defs = false
warn_return_any = false